fix: hierarchy under utils
