--- lib/data/hierarchy.ts
@@ -1,11 +1,0 @@
-export const HIERARCHY = [
-  'President',
-  'Vice President',
-  'Technical Head',
-  'Editorial Head',
-  'Corporate Head',
-  'Events Head',
-  'Lead',
-  'Associate Lead',
-  'Member',
-]
